Abstracted the model types into own files
