+++ src/app/tasks/task.model.ts
@@ -1,0 +1,7 @@
+export interface Task {
+  id: string;
+  userId: string;
+  title: string;
+  summary: string;
+  dueDate: string;
+}
