+++ src/app/tasks/task/task.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Task } from '../tasks.component';
+import { type  Task } from '../task.model';
 
 @Component({
   selector: 'app-task',
