+++ src/app/user/user.model.ts
@@ -1,0 +1,5 @@
+export interface User {
+  id: string;
+  name: string;
+  avatar: string;
+}
